refactor: admin component 구조 수정, ProductInfo 내 추가 컴포넌트 분리
