+++ src/refactoring/components/admin/index.ts
@@ -1,6 +1,9 @@
-export * from './CouponInfo';
-export * from './CouponManage';
-export * from './DiscountInfo';
-export * from './EditDiscountInfo';
-export * from './NewProduct';
-export * from './ProductInfo';
+export * from './couponManage/CouponInfo';
+export * from './couponManage/CouponManage';
+export * from './productManage/DiscountInfo';
+export * from './productManage/DiscountInfoEdit';
+export * from './productManage/NewProduct';
+export * from './productManage/ProductInfoDetail';
+export * from './productManage/ProductInfo';
+export * from './productManage/ProductInfoEdit';
+export * from './productManage/ProductInfoSummary';
